refactor: refatorado o componente DataPicker para DatePicker e ajustado o pop-up em telas menores, vindo com animação de baixo e ocupando toda a tela.
