--- src/components/DataPicker/DataPicker.scss
@@ -1,249 +1,0 @@
-.data-picker {
-  visibility: hidden;
-  opacity: 0;
-  transition: all 0.2s ease-in-out;
-  position: relative;
-  z-index: -22;
-
-  &.open {
-    visibility: visible;
-    opacity: 1;
-    z-index: 99;
-  }
-
-  .data-picker-calendar {
-    position: absolute;
-    top: 8px;
-    display: flex;
-    z-index: 99;
-    flex-direction: column;
-    gap: var(--s-spacing-small);
-    width: 288px;
-    background-color: var(--s-color-fill-default);
-    border: var(--s-border-width-hairline) solid var(--s-color-border-default);
-    border-radius: var(--s-border-radius-medium);
-    padding: var(--s-spacing-small);
-    box-shadow: var(--s-shadow-dropdown);
-
-    .data-picker-header {
-      display: flex;
-      justify-content: space-between;
-
-      .data-picker-header-date {
-        font: var(--s-typography-paragraph-regular);
-        color: var(--s-color-content-default);
-        background-color: var(--s-color-fill-default);
-        border-radius: var(--s-border-radius-small);
-        padding-left: var(--s-spacing-xx-small);
-        display: flex;
-        align-items: center;
-        cursor: pointer;
-
-        .button-icon {
-          transition: transform 0.3s ease;
-        }
-
-        &.open {
-          .button-icon {
-            transform: rotate(90deg);
-            transition: transform 0.3s ease;
-          }
-        }
-
-        &:hover {
-          background-color: var(--s-color-fill-default-hover);
-        }
-      }
-
-      .data-picker-header-buttons {
-        display: flex;
-        opacity: 1;
-        transition: all 0.2s ease-in-out;
-        visibility: visible;
-
-        &.hidden {
-          opacity: 0;
-          visibility: hidden;
-        }
-      }
-    }
-
-    .data-picker-body-container {
-      position: relative;
-      height: 224px;
-    }
-
-    .data-picker-body-months-years {
-      display: flex;
-      flex-direction: row;
-      position: absolute;
-      visibility: visible;
-      opacity: 1;
-      transition: all 0.2s ease-in-out;
-      width: 100%;
-      height: 100%;
-      align-items: center;
-      justify-content: center;
-
-      &.hidden {
-        visibility: hidden;
-        opacity: 0;
-      }
-
-      .data-picker-months {
-        display: flex;
-        height: 224px;
-        flex-direction: column;
-        flex: 1;
-        overflow-y: auto;
-        scrollbar-width: none;
-        -ms-overflow-style: none;
-        scroll-behavior: smooth;
-
-        .data-picker-months::-webkit-scrollbar {
-          display: none;
-        }
-
-        .months {
-          padding: var(--s-spacing-nano) var(--s-spacing-xx-small);
-          border-radius: var(--s-border-radius-medium);
-          background-color: var(--s-color-fill-default);
-          font: var(--s-typography-paragraph-regular);
-          color: var(--s-color-content-default);
-          cursor: pointer;
-
-          &:hover {
-            background-color: var(--s-color-fill-default-hover);
-          }
-
-          &.selected {
-            background-color: var(--s-color-fill-highlight);
-            color: var(--s-color-content-on-color);
-
-            &:hover {
-              background-color: var(--s-color-fill-highlight-hover);
-            }
-          }
-        }
-      }
-
-      .data-picker-years {
-        display: flex;
-        height: 224px;
-        flex-direction: column;
-        flex: 1;
-        overflow-y: auto;
-        scrollbar-width: none;
-        -ms-overflow-style: none;
-        scroll-behavior: smooth;
-
-        .data-picker-years::-webkit-scrollbar {
-          display: none;
-        }
-
-        .years {
-          padding: var(--s-spacing-nano) var(--s-spacing-xx-small);
-          border-radius: var(--s-border-radius-medium);
-          background-color: var(--s-color-fill-default);
-          font: var(--s-typography-paragraph-regular);
-          color: var(--s-color-content-default);
-          cursor: pointer;
-
-          &:hover {
-            background-color: var(--s-color-fill-default-hover);
-          }
-
-          &.selected {
-            background-color: var(--s-color-fill-highlight);
-            color: var(--s-color-content-on-color);
-
-            &:hover {
-              background-color: var(--s-color-fill-highlight-hover);
-            }
-          }
-        }
-      }
-    }
-
-    .data-picker-body {
-      display: flex;
-      flex-direction: column;
-      position: absolute;
-      visibility: visible;
-      opacity: 1;
-      transition: all 0.2s ease-in-out;
-
-      &.hidden {
-        visibility: hidden;
-        opacity: 0;
-      }
-
-      .data-picker-weekdays {
-        display: grid;
-        width: 256px;
-        grid-template-columns: repeat(7, 1fr);
-
-        .data-picker-week-day {
-          text-align: center;
-          font: var(--s-typography-caption-regular);
-          color: var(--s-color-content-light);
-          padding: var(--s-spacing-nano) var(--s-spacing-xx-small);
-        }
-      }
-
-      .data-picker-days {
-        display: grid;
-        width: 256px;
-        grid-template-columns: repeat(7, 1fr);
-
-        .data-picker-day {
-          background-color: var(--s-color-fill-default);
-          border-radius: var(--s-border-radius-small);
-          padding: var(--s-spacing-nano) var(--s-spacing-xx-small);
-          color: var(--s-color-content-default);
-          cursor: pointer;
-          text-align: center;
-          height: 32px;
-          font: var(--s-typography-paragraph-regular);
-          transition: all 0.2s ease-in-out;
-
-          &:hover {
-            background-color: var(--s-color-fill-default-hover);
-          }
-
-          &:disabled {
-            background-color: var(--s-color-fill-disable);
-          }
-
-          &.other-month {
-            cursor: auto;
-            color: var(--s-color-content-disable);
-
-            &:hover {
-              background-color: var(--s-color-fill-default);
-            }
-          }
-
-          &.selected {
-            background-color: var(--s-color-fill-highlight);
-            transition: all 0.2s ease-in-out;
-            color: var(--s-color-content-on-color);
-
-            &:hover {
-              background-color: var(--s-color-fill-highlight-hover);
-            }
-          }
-        }
-      }
-    }
-
-    .data-picker-footer {
-      display: flex;
-      justify-content: space-between;
-
-      .button {
-        width: fit-content;
-      }
-    }
-  }
-}
