+++ src/components/DatePicker/DatePicker.tsx
@@ -1,5 +1,5 @@
-import "./DataPicker.scss";
-import React, { useState } from "react";
+import "./DatePicker.scss";
+import React, { useState, useEffect } from "react";
 import Input from "../Input/Input";
 import { DatePickerProps } from "./DatePicker.types";
 import { useDateInput } from "./hooks/useDateInput";
@@ -12,6 +12,9 @@ import { DatePickerCalendar } from "./subcomponents/DatePickerCalendar";
  * Users can type a date directly (auto-masked as `DD/MM/YYYY`) or select one
  * from the visual calendar. The calendar supports month/year navigation and
  * a dedicated month+year picker panel for faster navigation.
+ *
+ * On mobile screens (< 425px), the calendar opens as a bottom sheet with a
+ * slide-up animation and full viewport width.
  *
  * @example
  * ```tsx
@@ -31,6 +34,7 @@ function DatePicker({
 }: DatePickerProps) {
   const [selectedDate, setSelectedDate] = useState<Date | null>(new Date());
   const [isCalendarOpen, setIsCalendarOpen] = useState(false);
+  const [isMobileScreen, setIsMobileScreen] = useState(false);
 
   const {
     inputValue,
@@ -45,6 +49,16 @@ function DatePicker({
     onDateChange: setSelectedDate,
   });
 
+  useEffect(() => {
+    const checkScreenSize = () => {
+      setIsMobileScreen(window.innerWidth < 425);
+    };
+
+    checkScreenSize();
+    window.addEventListener("resize", checkScreenSize);
+    return () => window.removeEventListener("resize", checkScreenSize);
+  }, []);
+
   const handleSelectDate = (date: Date) => {
     setSelectedDate(date);
     syncFromDate(date);
@@ -53,6 +67,10 @@ function DatePicker({
 
   const handleRevert = () => {
     handleSelectDate(new Date());
+  };
+
+  const handleCloseModal = () => {
+    setIsCalendarOpen(false);
   };
 
   return (
@@ -70,14 +88,35 @@ function DatePicker({
         disabled={disabled}
         skeleton={skeleton}
       />
-      <div className={`data-picker ${isCalendarOpen ? "open" : ""}`}>
-        <DatePickerCalendar
-          selectedDate={selectedDate}
-          onSelectDate={handleSelectDate}
-          onClose={() => setIsCalendarOpen(false)}
-          onRevert={handleRevert}
-        />
-      </div>
+
+      {!isMobileScreen && (
+        <div className={`data-picker ${isCalendarOpen ? "open" : ""}`}>
+          <DatePickerCalendar
+            selectedDate={selectedDate}
+            onSelectDate={handleSelectDate}
+            onClose={() => setIsCalendarOpen(false)}
+            onRevert={handleRevert}
+          />
+        </div>
+      )}
+
+      {isMobileScreen && (
+        <>
+          {isCalendarOpen && (
+            <div className="data-picker-modal-overlay" onClick={handleCloseModal} />
+          )}
+          <div className={`data-picker-modal ${isCalendarOpen ? "open" : ""}`}>
+            <div className="data-picker-modal-content">
+              <DatePickerCalendar
+                selectedDate={selectedDate}
+                onSelectDate={handleSelectDate}
+                onClose={handleCloseModal}
+                onRevert={handleRevert}
+              />
+            </div>
+          </div>
+        </>
+      )}
     </div>
   );
 }
