+++ src/components/DatePicker/utils/datePicker.utils.ts
@@ -1,6 +1,5 @@
 import { CalendarDay } from "../DatePicker.types";
 
-/** Full month names indexed 0–11. */
 export const MONTH_NAMES = [
   "January",
   "February",
@@ -16,7 +15,6 @@ export const MONTH_NAMES = [
   "December",
 ] as const;
 
-/** Abbreviated weekday labels used as column headers in the calendar grid. */
 export const WEEKDAY_LABELS = [
   "Sunday",
   "Monday",
@@ -27,15 +25,10 @@ export const WEEKDAY_LABELS = [
   "Saturday",
 ] as const;
 
-/** Total number of cells rendered in the calendar grid (6 rows × 7 columns). */
 const CALENDAR_GRID_SIZE = 42;
 
-/** Range of years available in the year picker, relative to the current year. */
 export const YEAR_RANGE = { past: 70, future: 25 } as const;
 
-/**
- * Formats a `Date` object as `DD/MM/YYYY`.
- */
 export function formatDate(date: Date): string {
   const day = String(date.getDate()).padStart(2, "0");
   const month = String(date.getMonth() + 1).padStart(2, "0");
@@ -43,10 +36,6 @@ export function formatDate(date: Date): string {
   return `${day}/${month}/${year}`;
 }
 
-/**
- * Parses a `DD/MM/YYYY` string into a `Date` object.
- * Returns `null` if the string is not a valid date.
- */
 export function parseDate(value: string): Date | null {
   const parts = value.split("/").map((part) => parseInt(part, 10));
   if (parts.length !== 3 || parts.some(isNaN)) return null;
@@ -62,10 +51,6 @@ export function parseDate(value: string): Date | null {
   return isValid ? date : null;
 }
 
-/**
- * Applies a raw digit string into a masked `DD/MM/YYYY` string.
- * Inserts slashes automatically at positions 2 and 4.
- */
 export function applyDateMask(digits: string): string {
   const d = digits.slice(0, 8);
   if (d.length > 4) return `${d.slice(0, 2)}/${d.slice(2, 4)}/${d.slice(4)}`;
@@ -73,10 +58,6 @@ export function applyDateMask(digits: string): string {
   return d;
 }
 
-/**
- * Builds the full 42-cell grid for a given month/year,
- * padding with days from the previous and next months as needed.
- */
 export function buildCalendarDays(year: number, month: number): CalendarDay[] {
   const firstDayOfMonth = new Date(year, month, 1).getDay();
   const daysInMonth = new Date(year, month + 1, 0).getDate();
@@ -98,10 +79,6 @@ export function buildCalendarDays(year: number, month: number): CalendarDay[] {
   return days;
 }
 
-/**
- * Clamps a day value so it stays valid after a month/year change.
- * Returns a new `Date` adjusted to the last day of the month if needed.
- */
 export function clampDayToMonth(
   year: number,
   month: number,
@@ -112,10 +89,6 @@ export function clampDayToMonth(
   return new Date(year, month, clampedDay);
 }
 
-/**
- * Generates the list of years available in the year picker.
- * Range spans from `currentYear - past` to `currentYear + future`.
- */
 export function buildYearRange(currentYear: number): number[] {
   return Array.from(
     { length: YEAR_RANGE.past + YEAR_RANGE.future + 1 },
